refactor(ui): move tailwind and postcss configs to ui package
- Move tailwind.config.ts, postcss.config.mjs, and globals.css to /ui package
- Move related package dependencies (tailwindcss, postcss, autoprefixer) from /next app to /ui package
- Move shadcn components.json config to /ui package
