+++ apps/next/postcss.config.mjs
@@ -1,8 +1,1 @@
-/** @type {import('postcss-load-config').Config} */
-const config = {
-  plugins: {
-    tailwindcss: {},
-  },
-};
-
-export default config;
+export { default } from "@repo/ui/postcss.config";
