+++ packages/ui/.eslintrc.js
@@ -3,10 +3,6 @@ module.exports = {
   root: true,
   extends: ["@repo/eslint-config/react-internal.js"],
   parser: "@typescript-eslint/parser",
-  parserOptions: {
-    project: "./tsconfig.lint.json",
-    tsconfigRootDir: __dirname,
-  },
   rules: {
     "no-redeclare": "off",
   },
